requirements.txt
aiohttp==3.6.2
alabaster==0.7.12
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.1
betterproto==1.2.5
black==19.10b0
certifi==2020.6.20
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
distlib==0.3.1
docutils==0.16
filelock==3.0.12
grpclib==0.3.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
httptools==0.1.1
hyperframe==5.2.0
identify==1.4.24
idna==2.10
imagesize==1.2.0
iniconfig==1.0.1
isort==4.3.21
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
multidict==4.7.6
nodeenv==1.4.0
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
pre-commit==2.6.0
py==1.9.0
pyasn1==0.4.8
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pytest==6.0.1
pytest-aiohttp==0.3.0
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
rsa==3.4.2
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
steamio==0.3.3
stringcase==1.2.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.10
uvicorn==0.11.6
uvloop==0.14.0
virtualenv==20.0.27
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
